rename the inventory package to diff
